+++ src/Synapse/Synapse/Models/DataPlaneModels/Artifact/Notebooks/PSNotebookCellOutputItem.cs
@@ -13,9 +13,7 @@
 // ----------------------------------------------------------------------------------
 
 using Azure.Analytics.Synapse.Artifacts.Models;
-using System;
-using System.Collections.Generic;
-using System.Text;
+using Newtonsoft.Json;
 
 namespace Microsoft.Azure.Commands.Synapse.Models
 {
@@ -31,16 +29,22 @@ namespace Microsoft.Azure.Commands.Synapse.Models
             this.Metadata = notebookCellOutputItem?.Metadata;
         }
 
+        [JsonProperty(PropertyName = "name")]
         public string Name { get; set; }
 
+        [JsonProperty(PropertyName = "execution_count")]
         public int? ExecutionCount { get; set; }
 
+        [JsonProperty(PropertyName = "output_type")]
         public string OutputType { get; set; }
 
+        [JsonProperty(PropertyName = "text")]
         public object Text { get; set; }
 
+        [JsonProperty(PropertyName = "data")]
         public object Data { get; set; }
 
+        [JsonProperty(PropertyName = "metadata")]
         public object Metadata { get; set; }
     }
 }