+++ src/Synapse/Synapse/Models/SynapseConstants.cs
@@ -197,6 +197,10 @@ namespace Microsoft.Azure.Commands.Synapse.Models
 
         public const string DataFlowDebugSessionPackage = nameof(DataFlowDebugSessionPackage);
 
+        public const string SqlScript = nameof(SqlScript);
+
+        public const int DefaultResultLimit = 5000;
+
         public static Dictionary<string, ComputeNodeSize> ComputeNodeSizes = new Dictionary<string, ComputeNodeSize>
         {
             {